requirements.txt
pip==23.0.1
pylance==0.3.7
APScheduler==3.10.0
mysql-connector-python==8.0.32
requests==2.28.2
load_dotenv==0.1.0
Flask==2.2.3
Werkzeug==2.2.3
click==8.1.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2